+++ c_api/src/wifi_network.cpp
@@ -1,5 +1,6 @@
 #include <wifi_pc_c_api/wifi_network.h>
 #include <wifi_pc/wifi_network.hpp>
+#include <stdint.h>
 #include <string>
 
 extern "C" {
@@ -15,10 +16,10 @@ const char* wpc_wifi_network_name(const WpcWifiNetwork* context) {
     return wifi_network->name().data();
 }
 
-int wpc_wifi_network_signal_quality(const WpcWifiNetwork* context) {
+uint8_t wpc_wifi_network_signal_quality(const WpcWifiNetwork* context) {
     auto wifi_network = (wpc::WifiNetwork*) context;
 
-    return wifi_network->signal_quality();
+    return (uint8_t) wifi_network->signal_quality();
 }
 
 } // extern "C"
